+++ packages/stylelint-config-wordpress/__tests__/properties.js
@@ -18,6 +18,7 @@ const invalidCss = (`
 	display: BLOCK;
 	margin-left: 20PX;
 	margin: 0;
+	argin: 20px;
 }
 `)
 /* eslint-enable */
@@ -44,12 +45,13 @@ test("There are warnings with invalid properties CSS", t => {
     const { errored, results } = data
     const { warnings } = results[0]
     t.truthy(errored, "errored")
-    t.is(warnings.length, 6, "flags six warnings")
+    t.is(warnings.length, 7, "flags seven warnings")
     t.is(warnings[0].text, "Expected \"#FFFFFF\" to be \"#ffffff\" (color-hex-case)", "correct warning text")
     t.is(warnings[1].text, "Expected \"#FFFFFF\" to be \"#FFF\" (color-hex-length)", "correct warning text")
     t.is(warnings[2].text, "Unexpected shorthand \"margin\" after \"margin-left\" (declaration-block-no-shorthand-property-overrides)", "correct warning text")
     t.is(warnings[3].text, "Expected single space after \":\" with a single-line declaration (declaration-colon-space-after)", "correct warning text")
-    t.is(warnings[4].text, "Expected \"PX\" to be \"px\" (unit-case)", "correct warning text")
-    t.is(warnings[5].text, "Expected \"BLOCK\" to be \"block\" (value-keyword-case)", "correct warning text")
+    t.is(warnings[4].text, "Unexpected unknown property \"argin\" (property-no-unknown)", "correct warning text")
+    t.is(warnings[5].text, "Expected \"PX\" to be \"px\" (unit-case)", "correct warning text")
+    t.is(warnings[6].text, "Expected \"BLOCK\" to be \"block\" (value-keyword-case)", "correct warning text")
   })
 })