+++ packages/stylelint-config-wordpress/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added: `selector-attribute-quotes` rule.
 * Added: `font-weight-notation` rule.
 * Added: `max-line-length` rule.
+* Added: `property-no-unknown` rule.
 * Deprecated `no-missing-eof-newline` rule. Use the new `no-missing-end-of-source-newline` rule instead.
 * Fixed `font-family-name-quotes` test warning message in `values.js`.